split auth workflow in separate tasks.
